kernel: Change UID to 64-bit
This is the same as symbian os, plus add check for get object

+++ src/emu/kernel/include/kernel/kernel_obj.h
@@ -32,7 +32,7 @@ namespace eka2l1 {
 
     namespace kernel {
         using handle = std::uint32_t;
-        using uid = std::uint32_t;
+        using uid = std::uint64_t;
 
         class process;
